# PrideHub-Backend-main/swagger.yml
openapi: 3.0.0
info:
  title: PrideHub API List
  description: This is a reference api list of PrideHub Backend.
  version: 1.0.0
  contact:
    email: yashmangal240@gmail.com
servers:
  - url: http://localhost:9999/api
    description: Development server
  - url: https://lgbtq-backend.onrender.com/api
    description: Production Server
paths:
...
  /comment/create/{postId}:
    post:
      tags:
        - Comment
      summary: Comment on a post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          description: Post ID
          schema:
            type: string
            format: ObjectId
      requestBody:
        $ref: '#/components/requestBodies/CreateComment'
      responses:
        '200':
          description: Comment Created Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /comment/fetch/{postId}:
    get:
      tags:
        - Comment
      summary: Get all comments of a post
      parameters:
        - in: path
          name: postId
          description: Post ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Comments Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /comment/delete/{commentId}:
    delete:
      tags:
        - Comment
      summary: Delete a comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentId
          description: Comment ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Comment Deleted Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /comment/like/{commentId}:
    put:
      tags:
        - Comment
      summary: Like a comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentId
          description: Comment ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Comment Liked Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /comment/unlike/{commentId}:
    put:
      tags:
        - Comment
      summary: Unlike a comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentId
          description: Comment ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Comment Unliked Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /comment/dislike/{commentId}:
    put:
      tags:
        - Comment
      summary: Dislike a comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentId
          description: Comment ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Comment Disliked Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /comment/undislike/{commentId}:
    put:
      tags:
        - Comment
      summary: Undislike a comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentId
          description: Comment ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Comment Undisliked Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    CreateCommentInput:
      type: object
      properties:
        comment:
          type: string
...
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    CreateComment:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateCommentInput'
        description: Create Comment
...